feat: created controller to handle upsert and deletion
